requirements.txt
numpy==1.21.2
matplotlib==3.4.3
Pillow==8.3.2
streamlit==0.88.0
opencv-python==4.5.3.56